Buffered writes to group price entries in files
Instead of writing price updates to disk immediately as singleton result
files they are now queued for up to FlushInterval (1 second) or
when MaxRecordsPerFile (10) items have been added to the buffer.

Instead of blocking on writes and responding with status 500 (but
discarding the actual error), data is now written in the background, and
errors are discarded with no indication that they occured. This will be
fixed when logging is integrated.

Flushing on shutdown is also not yet implemented.