Group price records in batched files
Implement the recordWriter interface according to the batching
requirements in the specification (limit number of records and duration
of batch).

New files are opened as necessary with increasing sequence numbers, and
closed when the batch is full.

Individual records are written immediately to maximize durability, but
files are only synced when a batch is full and the file becomes valid
JSON with the addition of the closing `]`.

To facilitate lookups keyed by productId, files are hardlinked in
corresponding directories, but this is only partly implemented (the
entry sequence number within these directories).

As per the examples, files are serialized as a JSON array with multiple
records and indentation, which is a bit hacky. A much cleaner/simpler
solution would be NDJSON, which would also have the advantage that
records that exceed the block size (which are possible because there are
no length constraints on the inputs) could be easily skipped by simply
ignoring all data in a file following the last newline, whereas with the
current format the last balanced closing brace within the top level
array has to be found, and a closing bracket appended to make the file
as a whole parseable.

This version still has an ugly sync.Mutex protecting updates to the
metadata fields in the filename struct embedded inthe batch struct. This
struct is modified by the linearizer's write loop thread on every new
record.  Finalization can happen either from a timer, or from this loop,
with the both possibilities coordinated by a sync.Once that starts a
goroutine that terminates, syncs and closes the file, and this goroutine
may race with writeRecord.