Truncate monotonic time & set expicit UTC timezone
time.Now() instantiates time.Time values in the floating time zone and
with monotonic clock information.

When serialized into JSON as RFC3339/ISO-8601 dates, the monotonic clock
data is discarded, and the time zone is assumed to be UTC.

This change makes it so that time data survives roundtripping, making
structs comparable in tests.