prevRequestListeners race condition workaround
Due to a flaw in the design of the linearizer loop, processing of new
price updates does not synchronously start loading previous price data
from disk, but does so by writing to a channel.

This reason for this is because the previous price fetching also being
required for last price requests when the data is not yet in memory, in
other words there is shared logic.

However, only the update request actually needs the result (the last
price request will be satisfied by the new value, but the previousPrice
field in the newly written data needs to have a consistent value), and
since updates are linearized into a total order, it only makes sense for
there to be one listener per product request.

However, due to the flaw an assumption required for this simplification is
broken - namely that once a price update request has started processing,
any last price request following it will not start loading previous
prices.

The workaround is simple and fairly efficient, just a goroutine that
passes the result of the read between the two channels. A proper fix
would be to refactor the previous price requests handling logic so that
registration of channels