[TECHDEBT] set entrySeq per productId correctly
More technical debt, in order to capture the entry sequence numbers per
product the previous sequence number must be recovered from the
preceding file.

Since this must be done synchronously, this is added to batchWriter
itself, and is technically redundant with with some of the previousPrice
loading, in that the directory will be scanned twice by the same
process.

Unfortunately this seems rather inherent in the specification
requirements (relying on plain files, and by extension POSIX, which does
not make any guarantees about ordering or consistency of directory
scanning) and I could not find a better solution to support pagination
of per product entries without deserializing and scanning through the
entire directory.