Decorator based priceModel wrapper adding PriceLog
A simple decorator that can delegates PriceLog queries by simply
embedding both interfaces.

For the main store implementation, the linearizer is responsible for the
main priceModel interface, using memStore for in memory price data, and
relying on priceLoader built on top of a snapshotFS to read entries from
persistent storage. This can now be wrapped with an auxillary
priceLoader (directly on top of the underlying readFS)), so that
storage.New() returns the extended price model.