Use filename metadata to perform sub-linear queries
Use the timestamp and entrySeq variables to handle paging more
efficiently.

Technically this is not sub-linear complexity, since directory
enumeration is O(N) and these need to be sorted, which is O(n log n).

However, this can be reasonably approached without unbounded memory use
by maintaing an LRU/ARC cache of directory listings, and ensuring that
writes to the filesystem are propagated to the file lists as well. With
such a mechanism in principle the file list can be accessed only once,
with all subsequent page requests on the same data requiring only
O(log N).